requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.3            # via telepot
apscheduler==3.6.3        # via -r requirements.in
asgiref==3.2.10           # via django
async-timeout==3.0.1      # via aiohttp
attrs==20.3.0             # via aiohttp
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via aiohttp, requests
django==3.1.2             # via -r requirements.in
google-api-core==1.22.4   # via google-api-python-client
google-api-python-client==1.12.3  # via -r requirements.in
google-auth-httplib2==0.0.4  # via -r requirements.in, google-api-python-client
google-auth-oauthlib==0.4.1  # via -r requirements.in
google-auth==1.22.1       # via google-api-core, google-api-python-client, google-auth-httplib2, google-auth-oauthlib
googleapis-common-protos==1.52.0  # via google-api-core
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests, yarl
multidict==5.0.2          # via aiohttp, yarl
oauthlib==3.1.0           # via requests-oauthlib
paho-mqtt==1.5.1          # via -r requirements.in
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pytz==2020.1              # via apscheduler, django, google-api-core, tzlocal
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.24.0          # via -r requirements.in, google-api-core, requests-oauthlib
rsa==4.6                  # via google-auth
six==1.15.0               # via apscheduler, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, protobuf
sqlparse==0.4.1           # via django
telepot==12.7             # via -r requirements.in
typing-extensions==3.7.4.3  # via aiohttp
tzlocal==2.1              # via apscheduler
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.10          # via requests, telepot
yarl==1.6.3               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
